# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: ISMS Builder API
  version: 1.0.0
  description: |
    REST API for the ISMS Builder platform — a self-hosted Information Security
    Management System (ISMS) according to ISO 27001:2022, NIS2, GDPR, BSI
    IT-Grundschutz and other frameworks.

    **Authentication:** All protected endpoints require a valid JWT cookie
    (`sm_session`). Obtain it via `POST /login`.

    **RBAC Roles (ascending rank):**
    - `reader` / `revision` (rank 1) — read access
    - `editor` / `dept_head` / `qmb` (rank 2) — create/edit
    - `contentowner` / `auditor` (rank 3) — approve, CISO/DSB settings
    - `admin` (rank 4) — user management, delete, system config

  contact:
    name: Claude Hecker
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
    description: Self-hosted instance
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sm_session
...
paths:
...
  /training:
    get:
      tags: [training]
      summary: List training records
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of training records
    post:
      tags: [training]
      summary: Create training record (editor+)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                category:
                  type: string
                targetGroup:
                  type: string
                dueDate:
                  type: string
                  format: date
                completionRate:
                  type: number
                linkedControls:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Created training
  /training/summary:
    get:
      tags: [training]
      summary: Training summary (completion rate, overdue)
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Training summary
  /training/{id}:
    get:
      tags: [training]
      summary: Get single training record
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Training record
    put:
      tags: [training]
      summary: Update training record (editor+)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
    delete:
      tags: [training]
      summary: Soft-delete training record (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /training/{id}/permanent:
    delete:
      tags: [training]
      summary: Permanently delete training record (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Permanently deleted
  /training/{id}/restore:
    post:
      tags: [training]
      summary: Restore soft-deleted training record (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Restored

  # ─── ADMIN ────────────────────────────────────────────────────────────────
...